Revert "Merge branch 'master' of https://github.com/linkerzhang/onnx into kezhan/add_function_private"
This reverts commit 711d41449eb96e58266d73e19e8b36be07b01971, reversing
changes made to 8bed1f974f9aebb77ab3eb42240d9e8975d8a222.

+++ onnx/backend/test/case/node/topk.py
@@ -29,12 +29,12 @@ class TopK(Base):
             [3, 2, 1],
             [7, 6, 5],
             [11, 10, 9],
-        ], dtype=np.float32)
+        ])
         indices_ref = np.array([
             [3, 2, 1],
             [3, 2, 1],
             [3, 2, 1],
-        ], dtype=np.int64)
+        ], dtype=np.int32)
 
         expect(node, inputs=[X], outputs=[values_ref, indices_ref],
                name='test_top_k')
